ultralytics-main-pose-fall/requirements.txt
shapely
numpy==1.24.4
opencv-python
einops
psutil
pyyaml
pandas
seaborn
tqdm
matplotlib
PyQt6==6.6.0
PyQt6-Qt6==6.6.0
PyQt6_sip==13.8.0
requests
efficientnet-pytorch
timm
mediapipe
